refactor project charter view, create
todo: update charter
