--- src/app/components/project-charter/create-project-charter-form/create-project-charter-form.component.html
@@ -1,168 +1,0 @@
-<mat-card class="p-6 max-w-4xl min-h-screen mx-auto">
-  @if (!isLoading) {
-  <!--------------------------------->
-  <!-- Create Project Charter Form -->
-  <!--------------------------------->
-
-  <form [formGroup]="projectCharterForm" (ngSubmit)="onSubmit()" class="max-w-4xl p-6 bg-white shadow-lg rounded-lg">
-    <div class="flex justify-between mb-6">
-      <div class="text-2xl md:text-4xl font-bold text-blue-600">TUYÊN NGÔN DỰ ÁN</div>
-      <span *ngIf="isCharterExist" nz-icon nzType="edit" nzTheme="outline" class="cursor-pointer" (click)="toggleEdit()"></span>
-    </div>
-
-    <!-- Project Overview Section -->
-    <section class="mb-8">
-      <div class="space-y-4">
-        <div>
-          <label class="block text-sm font-semibold bg-amber-100 p-2">BỐI CẢNH DỰ ÁN</label>
-          <textarea placeholder="Nhập bối cảnh dự án" formControlName="businessCase" class="p-2 bg-gray-50 border rounded w-full"></textarea>
-        </div>
-
-        <div>
-          <label class="block text-sm font-semibold bg-amber-100 p-2">MỤC TIÊU</label>
-          <textarea placeholder="Nhập mục tiêu" formControlName="goal" class="p-2 bg-gray-50 border rounded w-full"></textarea>
-        </div>
-
-        <div>
-          <label class="block text-sm font-semibold bg-amber-100 p-2">MỤC ĐÍCH</label>
-          <textarea placeholder="Nhập mục đích" formControlName="objective" class="p-2 bg-gray-50 border rounded w-full"></textarea>
-        </div>
-
-        <div>
-          <label class="block text-sm font-semibold bg-amber-100 p-2">PHẠM VI</label>
-          <textarea placeholder="Nhập phạm vi" formControlName="scope" class="p-2 bg-gray-50 border rounded w-full"></textarea>
-        </div>
-
-        <div>
-          <label class="block text-sm font-semibold bg-amber-100 p-2">HẠN CHẾ</label>
-          <textarea placeholder="Nhập hạn chế" formControlName="constraints" class="p-2 bg-gray-50 border rounded w-full"></textarea>
-        </div>
-
-        <div>
-          <label class="block text-sm font-semibold bg-amber-100 p-2">GIẢ ĐỊNH</label>
-          <textarea placeholder="Nhập giả định" formControlName="assumptions" class="p-2 bg-gray-50 border rounded w-full"></textarea>
-        </div>
-
-        <div>
-          <label class="block text-sm font-semibold bg-amber-100 p-2">KẾT QUẢ ĐẦU RA</label>
-          <textarea placeholder="Nhập kết quả đầu ra" formControlName="deliverables" class="p-2 bg-gray-50 border rounded w-full"></textarea>
-        </div>
-      </div>
-    </section>
-
-    <div *ngIf="isEditable && isCharterExist" class="text-right">
-      <button [disabled]="projectCharterForm.invalid" class="rounded-md" nz-button nzType="primary" type="button" (click)="saveEdit()">Save</button>
-    </div>
-
-    <div class="space-y-4 mt-5">
-      <div class="border-b border-gray-300 pb-4 mb-4"></div>
-
-      <!------------------------------------------------->
-      <!-- Create Many Phases ----------------------->
-      <!------------------------------------------------->
-
-      <p class="font-semibold text-xl text-gray-600 mb-2">GIAI ĐOẠN</p>
-      <div *ngIf="!isCharterExist" formArrayName="listCreatePhaseDtos">
-        @for (phase of listCreatePhaseDtos.controls; track $index; let i = $index) {
-        <div class="space-y-4 my-4" [formGroupName]="i">
-          <div class="flex justify-end">
-            <mat-icon (click)="showConfirmDelete(i)" class="text-right cursor-pointer text-red-500">delete</mat-icon>
-          </div>
-          <div class="sm:grid sm:grid-cols-4 items-center">
-            <label class="text-gray-600 pl-12">GIAI ĐOẠN</label>
-            <input placeholder="Nhập giai đoạn" formControlName="phaseName" type="text" class="col-span-3 rounded w-full" />
-          </div>
-          <!-- <div class="sm:grid sm:grid-cols-4 w-full gap-4">
-            <div class="sm:col-span-2 sm:grid sm:grid-cols-2 items-center">
-              <label class="col-span-1 text-gray-600 pl-12">NGÀY BẮT ĐẦU</label>
-              <nz-date-picker formControlName="startDate" nzPlaceHolder="Chọn ngày bắt đầu" class="rounded w-full ml-1" />
-            </div>
-            <div class="sm:col-span-2 sm:grid sm:grid-cols-2 items-center mt-4 sm:mt-0 sm:justify-end">
-              <label class="col-span-1 text-gray-600 pl-12">NGÀY KẾT THÚC</label>
-              <nz-date-picker formControlName="endDate" nzPlaceHolder="Chọn ngày kết thúc" class="rounded w-full" />
-            </div>
-          </div> -->
-          <div>
-            <div class="sm:grid sm:grid-cols-4 items-center">
-              <label class="col-span-1 text-gray-600 pl-12"
-                >NGÀY BẮT ĐẦU <br />
-                KẾT THÚC</label
-              >
-              <nz-range-picker [nzDisabledDate]="disabledDate" formControlName="startEndDate" class="col-span-3"></nz-range-picker>
-            </div>
-          </div>
-        </div>
-        }
-      </div>
-
-      <!------------------------------------------------->
-      <!------------- View Phase ------------------------>
-      <!------------------------------------------------->
-
-      <ng-container *ngIf="isCharterExist" class="mt-5">
-        <nz-table #nestedTable [nzShowPagination]="false" [nzData]="projectCharter?.phases!" [nzBordered]="true">
-          <thead>
-            <tr>
-              <th></th>
-              <th>Giai Đoạn</th>
-              <th>Ngày Bắt Đầu</th>
-              <th>Ngày Kết Thúc</th>
-            </tr>
-          </thead>
-          <tbody>
-            @for (data of nestedTable.data; track data) {
-            <tr [formGroupName]="$index">
-              <td [(nzExpand)]="data.expand"></td>
-              <td>{{ data.phaseName }}</td>
-              <td>{{ data.startDate | date : 'dd/MM/yyyy' }}</td>
-              <td>{{ data.endDate | date : 'dd/MM/yyyy' }}</td>
-            </tr>
-            <tr [nzExpand]="data.expand">
-              <nz-table #innerTable [nzData]="data.milestones" nzSize="small" [nzShowPagination]="false">
-                <thead>
-                  <tr>
-                    <th>Cột Mốc</th>
-                    <th>Miêu Tả</th>
-                    <th>Bắt Đầu</th>
-                    <th>Kết Thúc</th>
-                  </tr>
-                </thead>
-                <tbody>
-                  @for (data of innerTable.data; track data) {
-                  <tr>
-                    <td>{{ data.title }}</td>
-                    <td>{{ data.description }}</td>
-                    <td>{{ data.startDate | date : 'dd/MM/yyyy' }}</td>
-                    <td>{{ data.endDate | date : 'dd/MM/yyyy' }}</td>
-                  </tr>
-                  }
-                </tbody>
-              </nz-table>
-            </tr>
-            }
-          </tbody>
-        </nz-table>
-      </ng-container>
-
-      <div *ngIf="!isCharterExist" class="flex justify-end items-center gap-4">
-        <button
-          type="submit"
-          class="bg-orange-500 text-white p-2 rounded-md shadow-md"
-          [ngClass]="{
-            'bg-orange-500 text-white': projectCharterForm.valid,
-            'bg-gray-300 text-gray-500 cursor-not-allowed': projectCharterForm.invalid
-          }"
-          [disabled]="projectCharterForm.invalid"
-        >
-          Submit
-        </button>
-        <mat-icon (click)="addMilestone()" class="text-right cursor-pointer text-white bg-green-500 rounded-full shadow-sm hover:shadow-md" fontIcon="add">add</mat-icon>
-      </div>
-    </div>
-  </form>
-  } @else {
-  <div class="flex items-center justify-center">
-    <nz-spin nzSimple nzSize="large"></nz-spin>
-  </div>
-  }
-</mat-card>
