--- src/app/components/project-charter/create-project-charter-form/create-project-charter-form.component.scss
@@ -1,19 +1,0 @@
-textarea {
-  @apply p-2 border rounded w-full transition-colors duration-200;
-
-  &:disabled {
-    @apply bg-gray-50 cursor-not-allowed border-gray-300;
-  }
-
-  &:enabled {
-    @apply bg-white;
-
-    &:hover {
-      @apply border-gray-400;
-    }
-
-    &:focus {
-      @apply outline-none border-blue-500 ring-1 ring-blue-500;
-    }
-  }
-}
