+++ src/app/components/project-charter/project-detail/project-detail.component.html
@@ -1,0 +1,41 @@
+<nz-skeleton [nzLoading]="loading" [nzActive]="true" [nzTitle]="false" [nzParagraph]="{ rows: 12 }">
+  <div class="flex flex-col gap-5">
+    <nz-collapse>
+      <nz-collapse-panel [nzHeader]="'Thông tin chung dự án'" nzActive="true" class="bg-amber-200">
+        <div>
+          <nz-collapse>
+            <nz-collapse-panel [nzHeader]="'Tên dự án'" class="bg-amber-100">
+              <p>{{ projectGeneralInfo?.projectName }}</p>
+            </nz-collapse-panel>
+            <nz-collapse-panel [nzHeader]="'Miêu tả'" class="bg-amber-100">
+              <p>{{ projectGeneralInfo?.description }}</p>
+            </nz-collapse-panel>
+          </nz-collapse>
+        </div>
+      </nz-collapse-panel>
+    </nz-collapse>
+
+    <nz-collapse>
+      <nz-collapse-panel [nzHeader]="'Các thành viên dự án'" nzActive="true" class="bg-amber-200">
+        <nz-table #teamMemberTable [nzShowPagination]="false" nzBordered="true" [nzData]="projectGeneralInfo?.members ?? []">
+          <thead >
+            <tr>
+              <th>Tên Thành Viên</th>
+              <th>Vai Trò</th>
+              <th>Email</th>
+            </tr>
+          </thead>
+          <tbody>
+            @for (data of teamMemberTable.data; track $index) {
+            <tr>
+              <td>{{ data.fullName }}</td>
+              <td>{{ teamRoleLabels[data.roleInTeam] }}</td>
+              <td>{{ data.email }}</td>
+            </tr>
+            }
+          </tbody>
+        </nz-table>
+      </nz-collapse-panel>
+    </nz-collapse>
+  </div>
+</nz-skeleton>
