+++ src/app/components/project-pages/project-overview/project-charter/project-charter.component.html
@@ -1,109 +1,8 @@
 <div class="rounded-lg bg-white p-12 w-full">
   <h1 class="text-2xl font-bold text-blue-600 mb-6">TUYÊN NGÔN DỰ ÁN</h1>
 
-  <!-- Project Header Information -->
-  <div class="grid grid-cols-2 gap-4 mb-6">
-    <div>
-      <label class="text-sm font-semibold text-gray-600">TÊN DỰ ÁN</label>
-      <div class="p-2 bg-gray-50 border rounded">{{ projectOverview?.projectName }}</div>
-    </div>
-    <div>
-      <label class="text-sm font-semibold text-gray-600">CHỦ DỰ ÁN</label>
-      <div class="p-2 bg-gray-50 border rounded">{{ projectOverview?.leaderFullName }}</div>
-    </div>
+  <div class="flex gap-5 m-5 bg-white">
+    <app-project-detail class="w-1/2" [projectId]="projectOverview?.id!"></app-project-detail>
+    <app-project-charter-details class="w-1/2" [ProjectId]="projectOverview?.id!"></app-project-charter-details>
   </div>
-
-  <!-- Project Estimates -->
-  <div class="grid grid-cols-2 gap-4 mb-8">
-    <div>
-      <label class="text-sm font-semibold text-gray-600">NGÀY BẮT ĐẦU</label>
-      <div class="p-2 bg-gray-50 border rounded">{{ projectOverview?.startDate | date : 'dd-MM-yyyy' }}</div>
-    </div>
-    <div>
-      <label class="text-sm font-semibold text-gray-600">NGÀY KẾT THÚC (DỰ KIẾN)</label>
-      <div class="p-2 bg-gray-50 border rounded">{{ projectOverview?.endDate ? (projectOverview?.endDate | date : 'dd-MM-yyyy') : 'Chưa xác định' }}</div>
-    </div>
-  </div>
-
-  <!-- Project Overview -->
-  <section class="mb-8">
-    <h2 class="text-xl font-semibold mb-4">TỔNG QUAN DỰ ÁN</h2>
-    <div class="space-y-4">
-      <div>
-        <label class="block text-sm font-semibold bg-amber-100 p-2">GIẢ ĐỊNH</label>
-        <div class="p-2 bg-gray-50 border rounded">{{ projectOverview?.projectCharterResponseDto?.assumptions }}</div>
-      </div>
-      <div>
-        <label class="block text-sm font-semibold bg-amber-100 p-2">BỐI CẢNH DỰ ÁN</label>
-        <div class="p-2 bg-gray-50 border rounded">{{ projectOverview?.projectCharterResponseDto?.businessCase }}</div>
-      </div>
-      <div>
-        <label class="block text-sm font-semibold bg-amber-100 p-2">HẠN CHẾ</label>
-        <div class="p-2 bg-gray-50 border rounded">{{ projectOverview?.projectCharterResponseDto?.constraints }}</div>
-      </div>
-      <div>
-        <label class="block text-sm font-semibold bg-amber-100 p-2">MỤC TIÊU</label>
-        <div class="p-2 bg-gray-50 border rounded">{{ projectOverview?.projectCharterResponseDto?.goal }}</div>
-      </div>
-      <div>
-        <label class="block text-sm font-semibold bg-amber-100 p-2">KẾT QUẢ ĐẦU RA</label>
-        <div class="p-2 bg-gray-50 border rounded">{{ projectOverview?.projectCharterResponseDto?.deliverables }}</div>
-      </div>
-      <div>
-        <label class="block text-sm font-semibold bg-amber-100 p-2">PHẠM VI</label>
-        <div class="p-2 bg-gray-50 border rounded">{{ projectOverview?.projectCharterResponseDto?.scope }}</div>
-      </div>
-      <div>
-        <label class="block text-sm font-semibold bg-amber-100 p-2">MỤC ĐÍCH</label>
-        <div class="p-2 bg-gray-50 border rounded">{{ projectOverview?.projectCharterResponseDto?.objective }}</div>
-      </div>
-    </div>
-  </section>
-
-  <!-- Schedule -->
-  <section>
-    <h2 class="text-xl font-semibold mb-4">LỊCH TRÌNH DỰ ÁN</h2>
-    <nz-table #nestedTable class="w-full" [nzShowPagination]="false" [nzData]="projectOverview?.projectCharterResponseDto?.phases!" [nzBordered]="true">
-      <thead class="bg-gray-100">
-        <tr>
-          <th></th>
-          <th class="">TÊN GIAI ĐOẠN</th>
-          <th class=" ">NGÀY BẮT ĐẦU</th>
-          <th class="">NGÀY KẾT THÚC</th>
-        </tr>
-      </thead>
-      <tbody>
-        @for (data of nestedTable.data; track data) {
-        <tr>
-          <td [(nzExpand)]="data.expand"></td>
-          <td class="">{{ data.phaseName }}</td>
-          <td class="">{{ data.startDate | date : 'dd-MM-yyyy' }}</td>
-          <td class="">{{ data.endDate | date : 'dd-MM-yyyy' }}</td>
-        </tr>
-        <tr [nzExpand]="data.expand">
-          <nz-table #innerTable [nzData]="data.milestones" nzSize="small" [nzShowPagination]="false">
-            <thead>
-              <tr>
-                <th>Cột Mốc</th>
-                <th>Miêu Tả</th>
-                <th>Bắt Đầu</th>
-                <th>Kết Thúc</th>
-              </tr>
-            </thead>
-            <tbody>
-              @for (data of innerTable.data; track data) {
-              <tr>
-                <td>{{ data.title }}</td>
-                <td>{{ data.description }}</td>
-                <td>{{ data.startDate }}</td>
-                <td>{{ data.endDate }}</td>
-              </tr>
-              }
-            </tbody>
-          </nz-table>
-        </tr>
-        }
-      </tbody>
-    </nz-table>
-  </section>
 </div>
