+++ src/app/pages/project-charter-pages/create-project-charter/create-project-charter.page.ts
@@ -1,15 +1,20 @@
-import { Component } from '@angular/core';
-import { CreateProjectCharterFormComponent } from 'src/app/components/project-charter/create-project-charter-form/create-project-charter-form.component';
+import { Component, OnInit } from '@angular/core'
+import { ActivatedRoute } from '@angular/router'
+import { ProjectCharterDetailsComponent } from '../../../components/project-charter/project-charter-details/project-charter-details.component'
+import { ProjectDetailComponent } from 'src/app/components/project-charter/project-detail/project-detail.component'
+import { ViewTitleBarComponent } from 'src/app/layouts/view-title-bar/view-title-bar.component'
 
 @Component({
   selector: 'app-create-project-charter',
   templateUrl: './create-project-charter.page.html',
   styleUrls: ['./create-project-charter.page.scss'],
   standalone: true,
-  imports: [
-    CreateProjectCharterFormComponent
-  ],
+  imports: [ProjectCharterDetailsComponent, ProjectDetailComponent, ViewTitleBarComponent],
 })
-export class CreateProjectCharterPage {
-  constructor() {}
+export class CreateProjectCharterPage implements OnInit {
+  projectId = ''
+  constructor(private activeRoute: ActivatedRoute) {}
+  ngOnInit(): void {
+    this.projectId = this.activeRoute.parent?.snapshot.paramMap.get('id')!
+  }
 }
