+++ src/app/services/project-charter.service.ts
@@ -3,6 +3,8 @@ import { Injectable } from '@angular/core'
 import { ApplicationConfigService } from '../core/config/application-config.service'
 import { ProjectCharterFormModel } from '../shared/models/project-charter/project-charter-create.model'
 import { ProjectCharterUpdateModel } from '../shared/models/project-charter/project-charter-update.model'
+import { BehaviorSubject, Observable, tap } from 'rxjs'
+import { ProjectGeneralInformationModel } from '../shared/models/project/project-general-info.model'
 
 @Injectable({
   providedIn: 'root',
@@ -10,9 +12,15 @@ import { ProjectCharterUpdateModel } from '../shared/models/project-charter/proj
 export class ProjectCharterService {
   constructor(private http: HttpClient, private applicationConfigService: ApplicationConfigService) {}
 
+  refreshCharter$ = new BehaviorSubject<boolean>(true)
+
   create(projectId: string, projectCharter: ProjectCharterFormModel) {
     const url = `/api/projects/${projectId}/project-charters`
-    return this.http.post(this.applicationConfigService.getEndpointFor(url), projectCharter)
+    return this.http.post(this.applicationConfigService.getEndpointFor(url), projectCharter).pipe(
+      tap(() => {
+        this.refreshCharter$.next(true)
+      })
+    )
   }
 
   get(projectId: string) {
@@ -20,8 +28,17 @@ export class ProjectCharterService {
     return this.http.get(this.applicationConfigService.getEndpointFor(url))
   }
 
-  edit(projectId: string, projectCharter:ProjectCharterUpdateModel) {
+  edit(projectId: string, projectCharter: ProjectCharterUpdateModel) {
     const url = `/api/projects/${projectId}/project-charters`
-    return this.http.put(this.applicationConfigService.getEndpointFor(url), projectCharter)
+    return this.http.put(this.applicationConfigService.getEndpointFor(url), projectCharter).pipe(
+      tap(() => {
+        this.refreshCharter$.next(true)
+      })
+    )
+  }
+
+  getProjectGeneralInfo(projectId: string): Observable<ProjectGeneralInformationModel> {
+    const url = `/api/projects/${projectId}/general-information`
+    return this.http.get<ProjectGeneralInformationModel>(this.applicationConfigService.getEndpointFor(url))
   }
 }
