+++ src/app/shared/models/project/project-general-info.model.ts
@@ -1,0 +1,9 @@
+import { TeamMemberModel } from '../user/team-member.model'
+
+export type ProjectGeneralInformationModel = {
+  projectName: string
+  description: string
+  logoUrl: string
+  members: TeamMemberModel[]
+  id: string
+}
